fix: meat/get in front
