+++ test-web/src/components/DataListView/PADataListComp.js
@@ -6,7 +6,7 @@ import Spinner from 'react-bootstrap/Spinner';
 import { usePredictedMeatListFetch } from '../../API/getPredictedMeatListSWR';
 
 // 데이터 예측 페이지 목록 컴포넌트
-const PADataListComp = ({ startDate, endDate, pageOffset }) => {
+const PADataListComp = ({ startDate, endDate, pageOffset, specieValue }) => {
   // 고기 데이터 목록
   const [meatList, setMeatList] = useState([]);
   // 데이터 전체 개수
@@ -15,12 +15,7 @@ const PADataListComp = ({ startDate, endDate, pageOffset }) => {
   const [currentPage, setCurrentPage] = useState(1);
   // 한페이지당 보여줄 개수
   const [count, setCount] = useState(5);
-  const [specieValue, setSpecieValue] = useState('전체')
-
-  const handleSpeciesChange = (event) => {
-    setSpecieValue(event.target.value);
-  };
-
+  
 
   // API fetch 데이터 전처리
   const processPAMeatDatas = (data) => {
@@ -103,17 +98,6 @@ const PADataListComp = ({ startDate, endDate, pageOffset }) => {
   // 정상 데이터 로드 된 경우
   return (
     <div>
-       <Select
-        labelId="species"
-        id="species"
-        value={specieValue}
-        onChange={handleSpeciesChange}
-        label="종류"
-      >
-        <MenuItem value="전체">전체</MenuItem>
-        <MenuItem value="소">소</MenuItem>
-        <MenuItem value="돼지">돼지</MenuItem>
-      </Select>
       <div style={style.listContainer}>
         {meatList !== undefined && (
           <DataList
