+++ test-web/src/components/DataListView/RejectedDataListComp.js
@@ -12,6 +12,7 @@ const RejectedDataListComp = ({
   startDate, // 조회 시작 날짜
   endDate, // 조회 종료 날짜
   pageOffset, // 조회 페이지 offset
+  specieValue,
 }) => {
   // 고기 데이터 목록
   const [meatList, setMeatList] = useState([]);
@@ -22,7 +23,6 @@ const RejectedDataListComp = ({
   // 한페이지당 보여줄 개수
   const [count, setCount] = useState(5);
   const totalPages = Math.ceil(totalData / count);
-  const [specieValue, setSpecieValue] = useState('전체');
 
   // API fetch 데이터 전처리
   const processRejectedMeatDatas = (data) => {
@@ -31,9 +31,7 @@ const RejectedDataListComp = ({
     // 반려데이터
     setMeatList(data['반려']);
   };
-  const handleSpeciesChange = (event) => {
-    setSpecieValue(event.target.value);
-  };
+
   //API fetch
   const { data, isLoading, isError } = useRejectedMeatListFetch(
     currentPage - 1,
@@ -89,17 +87,6 @@ const RejectedDataListComp = ({
   // 정상 데이터 로드 된 경우
   return (
     <div>
-      <Select
-        labelId="species"
-        id="species"
-        value={specieValue}
-        onChange={handleSpeciesChange}
-        label="종류"
-      >
-        <MenuItem value="전체">전체</MenuItem>
-        <MenuItem value="소">소</MenuItem>
-        <MenuItem value="돼지">돼지</MenuItem>
-      </Select>
       <div style={style.listContainer}>
         {meatList !== undefined && (
           <DataList
