+++ test-web/src/components/DataListView/excelExport.js
@@ -2,9 +2,12 @@ import * as XLSX from 'xlsx';
 import { apiIP } from '../../config';
 
 // export 할 육류 데이터 목록 fetch
-const getDataListJSON = async () => {
+const getDataListJSON = async ({ startDate, endDate, specieValue }) => {
+  console.log('???:');
   try {
-    const response = await fetch(`http://${apiIP}/meat/get`);
+    const response = await fetch(
+      `http://${apiIP}/meat/get?start=${startDate}&end=${endDate}&specieValue=${specieValue}`
+    );
     if (!response.ok) {
       throw new Error('Network response was not ok');
     }
@@ -28,6 +31,7 @@ const downloadExcel = (data) => {
 // json 데이터 가공
 const DataListJSON2Excel = (rawData) => {
   let newData = [];
+  console.log(rawData);
 
   // id_list에 있는 ID들을 사용하여 meat_dict에서 데이터를 가져옴
   rawData.id_list.forEach((id) => {
